+++ uni-stark-ext/src/lib.rs
@@ -6,7 +6,7 @@ extern crate alloc;
 
 mod config;
 mod folder;
-mod interaction_air_builder;
+mod interaction;
 mod keygen;
 mod proof;
 mod prover;
@@ -14,7 +14,7 @@ mod verifier;
 
 pub use config::*;
 pub use folder::*;
-pub use interaction_air_builder::*;
+pub(crate) use interaction::*;
 pub use keygen::*;
 pub use p3_air_ext::*;
 pub use proof::*;
